backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Database
supabase==2.5.0

# AI/ML
anthropic==0.34.0
lightgbm==4.4.0
scikit-learn==1.5.0
numpy==1.26.4
pandas==2.2.2
shap==0.46.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Utilities
python-dotenv==1.0.0
pydantic==2.8.0
pydantic-settings==2.4.0
httpx==0.27.0